+++ conf/script/src/build_system/cmd/setup/cli/cli_print_meson_cmd.py
@@ -5,7 +5,7 @@ from build_system.cmd.hierarchy.consts import BUILD_SYSTEM_NAME
 
 
 def print_meson_main_file(meson_main_file: Path) -> None:
-    from build_system.cmd.setup.cli.cli_color import colorize_label, colorize_path
+    from build_system.cmd.setup.cli.colorize import colorize_label, colorize_path
 
     meson_cmd_label: Final[str] = colorize_label(label=f'{BUILD_SYSTEM_NAME.capitalize()} main file') + ': '
 
@@ -16,7 +16,7 @@ def print_meson_main_file(meson_main_file: Path) -> None:
 
 
 def print_meson_cmd(meson_cli_args: list[str]) -> None:
-    from build_system.cmd.setup.cli.cli_color import colorize_label, colorize_path
+    from build_system.cmd.setup.cli.colorize import colorize_label, colorize_path
 
     meson_cmd_label: Final[str] = colorize_label(label=f'{BUILD_SYSTEM_NAME.capitalize()} cmd') + ': '
