+++ conf/script/src/build_system/cmd/setup/cli/colorize.py
